Move ArgparseCliParser unit tests to own test module
