+++ tests/OTAnalytics/plugin_ui/test_cli.py
@@ -1,9 +1,8 @@
-import sys
 from datetime import datetime
 from pathlib import Path
 from shutil import copy2, rmtree
 from typing import Any
-from unittest.mock import Mock, call, patch
+from unittest.mock import Mock, call
 
 import pytest
 
@@ -78,7 +77,6 @@ from OTAnalytics.plugin_intersect.simple_intersect import (
 from OTAnalytics.plugin_intersect_parallelization.multiprocessing import (
     MultiprocessingIntersectParallelization,
 )
-from OTAnalytics.plugin_parser.argparse_cli_parser import ArgparseCliParser
 from OTAnalytics.plugin_parser.export import (
     AddSectionInformationExporterFactory,
     FillZerosExporterFactory,
@@ -92,11 +90,9 @@ from OTAnalytics.plugin_parser.otvision_parser import (
 )
 from OTAnalytics.plugin_prototypes.eventlist_exporter.eventlist_exporter import (
     AVAILABLE_EVENTLIST_EXPORTERS,
-    OTC_CSV_FORMAT_NAME,
     OTC_OTEVENTS_FORMAT_NAME,
 )
 from OTAnalytics.plugin_ui.cli import (
-    EventFormat,
     InvalidSectionFileType,
     OTAnalyticsCli,
     SectionsFileDoesNotExist,
@@ -177,55 +173,6 @@ def create_cli_args(
         logfile,
         logfile_overwrite,
     )
-
-
-class TestCliArgumentParser:
-    def test_parse_with_valid_cli_args(self) -> None:
-        track_file_1 = f"track_file_1.{DEFAULT_TRACK_FILE_TYPE}"
-        track_file_2 = f"track_file_2.{DEFAULT_TRACK_FILE_TYPE}"
-        sections_file = "section_file.otflow"
-        save_name = "stem"
-        save_suffix = "suffix"
-        log_file = "path/to/my_log.log"
-
-        cli_args: list[str] = [
-            "path",
-            "--cli",
-            "--ottrks",
-            track_file_1,
-            track_file_2,
-            "--otflow",
-            sections_file,
-            "--save-name",
-            save_name,
-            "--save-suffix",
-            save_suffix,
-            "--event-format",
-            EventFormat.CSV.value,
-            "--count-interval",
-            "15",
-            "--num-processes",
-            "3",
-            "--logfile",
-            log_file,
-            "--logfile_overwrite",
-        ]
-        with patch.object(sys, "argv", cli_args):
-            parser = ArgparseCliParser()
-            args = parser.parse()
-            assert args == CliArguments(
-                True,
-                False,
-                [track_file_1, track_file_2],
-                sections_file,
-                save_name,
-                save_suffix,
-                AVAILABLE_EVENTLIST_EXPORTERS[OTC_CSV_FORMAT_NAME],
-                15,
-                3,
-                log_file,
-                True,
-            )
 
 
 class TestOTAnalyticsCli:
